+++ lib/Phpactor.php
@@ -9,6 +9,7 @@ use Phpactor\Extension\Behat\BehatExtension;
 use Phpactor\Extension\Behat\BehatSuggestExtension;
 use Phpactor\Extension\ComposerInspector\ComposerInspectorExtension;
 use Phpactor\Extension\Configuration\ConfigurationExtension;
+use Phpactor\Extension\Core\Trust\Trust;
 use Phpactor\Extension\Debug\DebugExtension;
 use Phpactor\Extension\LanguageServerBlackfire\LanguageServerBlackfireExtension;
 use Phpactor\Extension\LanguageServerConfiguration\LanguageServerConfigurationExtension;
@@ -46,6 +47,7 @@ use Phpactor\Extension\WorseReflectionAnalyse\WorseReflectionAnalyseExtension;
 use Phpactor\Indexer\Extension\IndexerExtension;
 use RuntimeException;
 use Symfony\Component\Console\Output\ConsoleOutputInterface;
+use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Phpactor\Container\PhpactorContainer;
 use Phpactor\Extension\Core\CoreExtension;
@@ -79,6 +81,7 @@ use Phpactor\ConfigLoader\ConfigLoaderBuilder;
 use Phpactor\Extension\ReferenceFinderRpc\ReferenceFinderRpcExtension;
 use Phpactor\Extension\ReferenceFinder\ReferenceFinderExtension;
 use Symfony\Component\Filesystem\Path;
+use XdgBaseDir\Xdg;
 use function ini_set;
 use function sprintf;
 
@@ -86,6 +89,7 @@ class Phpactor
 {
     public static function boot(InputInterface $input, OutputInterface $output, string $vendorDir, ?string $phpactorBin = null): Container
     {
+        $errorOutput = $output instanceof ConsoleOutputInterface ? $output->getErrorOutput() : new NullOutput();
         $config = [];
 
         $projectRoot = getcwd();
@@ -94,19 +98,42 @@ class Phpactor
             $projectRoot = $input->getParameterOption([ '--working-dir', '-d' ]);
         }
 
+        if (!is_string($projectRoot)) {
+            throw new RuntimeException(sprintf(
+                'Unexpected type for project root, expected string got: %s',
+                get_debug_type($projectRoot)
+            ));
+        }
+
         $commandName = $input->getFirstArgument();
+
+        $trustPath = (new Xdg())->getHomeDataDir() . '/phpactor/trust.json';
+        $trust = Trust::load($trustPath);
 
         $loader = ConfigLoaderBuilder::create()
             ->enableJsonDeserializer('json')
             ->enableYamlDeserializer('yaml')
             ->addXdgCandidate('phpactor', 'phpactor.json', 'json')
-            ->addXdgCandidate('phpactor', 'phpactor.yml', 'yaml')
-            ->addCandidate($projectRoot . '/.phpactor.json', 'json')
-            ->addCandidate($projectRoot . '/.phpactor.yml', 'yaml')
-            ->loader();
+            ->addXdgCandidate('phpactor', 'phpactor.yml', 'yaml');
 
+        $projectCandidates = [
+            [$projectRoot . '/.phpactor.json', 'json'],
+            [$projectRoot . '/.phpactor.yml', 'yaml'],
+        ];
+
+        if ($trust->isTrusted($projectRoot) === true) {
+            foreach ($projectCandidates as [$path, $type]) {
+                $loader = $loader->addCandidate($path, $type);
+            }
+        }
+
+        $loader = $loader->loader();
         $config = $loader->load();
         $config[CoreExtension::PARAM_COMMAND] = $input->getFirstArgument();
+        $config[CoreExtension::PARAM_PROJECT_CONFIG_CANDIDATES] = array_column($projectCandidates, 0);
+
+        $config[CoreExtension::PARAM_TRUST] = $trust;
+
         if ($phpactorBin) {
             $config[LanguageServerExtension::PARAM_PHPACTOR_BIN] = $phpactorBin;
         }
@@ -142,6 +169,19 @@ class Phpactor
             $xdebug->check();
             unset($xdebug);
         }
+
+        if (!$trust->isTrusted($projectRoot)) {
+            foreach ($projectCandidates as [$candidate, $_]) {
+                if (file_exists($candidate)) {
+                    $errorOutput->writeln(sprintf(
+                        '<fg=yellow>Local config "%s" found but it\'s in an untrusted directory, ' .
+                        'run `phpactor config:trust` if you want it to be loaded</>',
+                        basename($candidate),
+                    ));
+                }
+            }
+        }
+
 
         $extensionNames = [
             CoreExtension::class,
@@ -231,9 +271,7 @@ class Phpactor
             $schema = new Resolver();
 
             if (!class_exists($extensionClass)) {
-                if ($output instanceof ConsoleOutputInterface) {
-                    $output->getErrorOutput()->writeln(sprintf('<error>Extension "%s" does not exist</>', $extensionClass). "\n");
-                }
+                $errorOutput->writeln(sprintf('<error>Extension "%s" does not exist</>', $extensionClass). "\n");
                 continue;
             }